FEATURE[35] - Inicia adicao do BLOC na pagina de login
