+++ lib/routing/router.dart
@@ -1,7 +1,7 @@
 import 'package:cafe_pra_ja/routing/app_shell.dart';
 import 'package:cafe_pra_ja/routing/routes.dart';
 import 'package:cafe_pra_ja/ui/auth/cadastro/cadastro_screen.dart';
-import 'package:cafe_pra_ja/ui/auth/login/login_screen.dart';
+import 'package:cafe_pra_ja/ui/auth/login/login_screen_page.dart';
 import 'package:cafe_pra_ja/ui/bemvindo/bemvindo_screeen.dart';
 import 'package:cafe_pra_ja/ui/checkout/checkout_screen.dart';
 import 'package:cafe_pra_ja/ui/core/localization/cafe_string.dart';
