+++ lib/ui/_screen_page.dart
@@ -21,6 +21,10 @@ class _State extends State<Subject> {
   void initState() {
     super.initState();
   }
+  /// Metodos
+  void m() {
+    
+  }
   /// Widget
 
   Widget _page(state) {
